+++ src/main/java/com/search/instagramsearching/jwt/util/TokenProperties.java
@@ -9,8 +9,8 @@ public interface TokenProperties {
     String INVALID = "INVALID";
     String EXPIRED = "EXPIRED";
     // Access JWT 토큰의 유효기간: 30분 (단위: milliseconds)
-    int ACCESS_TOKEN_VALID_TIME = 30 * 60 * 1000;
+    long ACCESS_TOKEN_VALID_TIME = 30 * 60 * 1000;
 
     // Refresh JWT 토큰의 유효기간: 하루 (단위: milliseconds)
-    int REFRESH_TOKEN_VALID_TIME = 24 * 60 * 60 * 1000;
+    long REFRESH_TOKEN_VALID_TIME = 24 * 60 * 60 * 1000;
 }
